# fix: cache apiUsageMonitoring filter creation based on args