fix: creation of apiUsageMonitoring filter is costly so cache filter creation based on args

```
% benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/zalando/skipper/filters/apiusagemonitoring cpu: AMD Ryzen 7 PRO 4750U with Radeon Graphics
                                           │     old.txt     │               new.txt               │
                                           │     sec/op      │   sec/op     vs base                │
_CreateFilter_FullConfigSingleApiNakadi-16   452971.0n ± 36%   670.7n ± 1%  -99.85% (p=0.000 n=10)

                                           │    old.txt    │               new.txt                │
                                           │     B/op      │     B/op      vs base                │
_CreateFilter_FullConfigSingleApiNakadi-16   49.322Ki ± 0%   1.125Ki ± 0%  -97.72% (p=0.000 n=10)

                                           │    old.txt    │              new.txt               │
                                           │   allocs/op   │ allocs/op   vs base                │
_CreateFilter_FullConfigSingleApiNakadi-16   1379.000 ± 0%   1.000 ± 0%  -99.93% (p=0.000 n=10)
```